+++ src/pages/home/Home.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Cards } from '../../components/cards/Cards';
+import { Cards } from '../../components/Cards111/Cards';
 import { Search } from '../../components/search/Search';
 import './home.css';
